10-Authentification : End

+++ app/Concern/Admin/TraitAdminController.php
@@ -3,11 +3,15 @@
 namespace App\Concern\Admin;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 /**
  * Trait TraitAdminController
  *
  * Basic CRUD controller
+ *
+ * ## L'idée de base étant d'avoir un Trait qui reprend le CRUD.
+ * ## Ainsi il suffira soit de ne pas toucher soit de réécrire les fonctions pour les controllers particuliés.
  *
  * @package App\Concern\Admin
  */
@@ -28,6 +32,9 @@ trait TraitAdminController
      */
     public function index()
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $items = $this->model::all();
         return view('admin.' . $this->view . '.index', compact('items'));
     }
@@ -39,6 +46,9 @@ trait TraitAdminController
      */
     public function create()
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $item = new $this->model;
         return view('admin.' . $this->view . '.form', compact('item'));
     }
@@ -51,6 +61,9 @@ trait TraitAdminController
      */
     public function store(Request $request)
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $this->model::create($request->all());
         return $this->index()->with('success', 'Nouvel entité créé');
     }
@@ -63,6 +76,9 @@ trait TraitAdminController
      */
     public function show($id)
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $item = $this->model::where('id', $id)->first();
         return view('admin.' . $this->view . '.show', compact('item'));
     }
@@ -75,6 +91,9 @@ trait TraitAdminController
      */
     public function edit($id)
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $item = $this->model::where('id', $id)->first();
         return view('admin.' . $this->view . '.form', compact('item'));
     }
@@ -88,6 +107,9 @@ trait TraitAdminController
      */
     public function update(Request $request, $id)
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $item = $this->model::where('id', $id)->first();
         $item->update($request->all());
         return $this->index()->with('success', 'Entité modifié avec succès');
@@ -101,6 +123,9 @@ trait TraitAdminController
      */
     public function destroy($id)
     {
+        if (! Auth::user()->hasRole('admin')) {
+            return redirect()->back();
+        }
         $item = $this->model::where('id', $id)->first();
         $item->delete();
         return $this->index()->with('success', 'Entité supprimé');
